// types/application.ts
export interface Application {
    id: string;
    company: string
    url: string | ""
    jobTitle: string
    requirements: string | ""
    appliedThrough: string
    salary: string | ""
    locationType: string
    location: string
    dateApplied: string
    status: string
    latestReply: string | ""
    interviewDate: string | ""
    feedback: string | ""
  }